Split mod.rs into root, parse, load submodules
Reduces src/config/mod.rs from 965 lines to a 53-line barrel module by
splitting the remaining Config struct, TOML parsing, and file/env loading
into three sibling modules. mod.rs now only contains module declarations,
re-exports, and the shared default_true() helper.

- root.rs: Config struct, Default impl, methods, 5 tests
- parse.rs: parse_config_with_defaults + merge_toml_values, 8 tests
- load.rs: load_config, save_config, parse_bool_env, 1 test

Public API unchanged: crate::config::{Config, load_config, save_config,
parse_config_with_defaults} all still resolve through re-exports.
